pyproject.toml
[project]
name = "yaml-fail-repro"
version = "0.1.0"
description = "validate if outputschema used in yaml in maf"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework-core>=1.0.0b260128",
    "agent-framework-azure-ai>=1.0.0b260128",
    "agent-framework-devui>=1.0.0b260128",
    "agent-framework-declarative>=1.0.0b260128",
    "python-dotenv>=1.0.0",
    "pydantic>=2,<3",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-api>=1.39.1",
]

[tool.uv]
prerelease = "allow"
